+++ design-system/tokens/src/tokens/atlassian-light/color/border.tsx
@@ -6,6 +6,9 @@ const color: ValueSchema<BorderColorTokenSchema> = {
       '[default]': {
         value: 'N300A',
       },
+      bold: {
+        value: 'N600',
+      },
       inverse: {
         value: 'N0',
       },