+++ linking-platform/link-provider/src/reducers/index.ts
@@ -8,53 +8,87 @@ import {
   ACTION_ERROR_FALLBACK,
   CardStore,
   CardState,
+  CardActionType,
+  ACTION_RELOADING,
+  CardAction,
 } from '@atlaskit/linking-common';
-import { CardReducerMap, CardReducer } from '../types';
+import { CardReducer } from '../types';
 import { getStatus } from '../helpers';
+import { AnyAction } from 'redux';
 
-const cardReducerMap: CardReducerMap<CardState, JsonLd.Response> = {
-  [ACTION_PENDING]: (_state, { type }) => {
-    return { status: type };
-  },
-  [ACTION_RESOLVING]: (state, { type }) => {
-    return { ...state, status: type };
-  },
-  [ACTION_RESOLVED]: (state, { type, payload }) => {
-    const nextDetails = payload;
-    const nextState = { ...state };
-    if (nextDetails) {
-      nextState.status = getStatus(nextDetails);
-      nextState.details = nextDetails;
-    } else {
-      // Keep the pre-existing data in the store. If there
-      // is no data, the UI should handle this gracefully.
-      nextState.status = type;
-    }
-    return nextState;
-  },
-  [ACTION_ERROR]: (state, { type, error }) => {
-    return { ...state, status: type, error };
-  },
-  [ACTION_ERROR_FALLBACK]: (state, { type, payload, error }) => {
-    return { ...state, status: type, error, details: payload };
-  },
+const isCardAction = (action: AnyAction): action is CardAction => {
+  return [
+    ACTION_PENDING,
+    ACTION_RESOLVING,
+    ACTION_RESOLVED,
+    ACTION_RELOADING,
+    ACTION_ERROR,
+    ACTION_ERROR_FALLBACK,
+  ].includes(action.type);
 };
+
+function persistPayloadToState(
+  payload: JsonLd.Response<JsonLd.Data.BaseData> | undefined,
+  state: CardState,
+  type: Exclude<CardActionType, 'reloading'>,
+) {
+  const nextDetails = payload;
+  const nextState = { ...state };
+  if (nextDetails) {
+    nextState.status = getStatus(nextDetails);
+    nextState.details = nextDetails;
+  } else {
+    // Keep the pre-existing data in the store. If there
+    // is no data, the UI should handle this gracefully.
+    nextState.status = type;
+  }
+  return nextState;
+}
+
+function reduceToNextState(state: CardState, action: CardAction) {
+  switch (action.type) {
+    case ACTION_PENDING: {
+      return { status: action.type };
+    }
+    case ACTION_RESOLVING: {
+      return { ...state, status: action.type };
+    }
+    case ACTION_RESOLVED: {
+      return persistPayloadToState(action.payload, state, action.type);
+    }
+    case ACTION_RELOADING: {
+      return persistPayloadToState(action.payload, state, 'resolved');
+    }
+    case ACTION_ERROR: {
+      return { ...state, status: action.type, error: action.error };
+    }
+    case ACTION_ERROR_FALLBACK: {
+      return {
+        ...state,
+        status: action.type,
+        error: action.error,
+        details: action.payload,
+      };
+    }
+  }
+}
 
 export const cardReducer: CardReducer<CardStore, JsonLd.Response> = (
   state,
   action,
 ) => {
-  if (cardReducerMap[action.type]) {
-    const cardState = state[action.url];
-    // Card may have reached the same state on account of multiple of the same
-    // URL being present in an editor session. E.g. page with N links to one resource.
-    const hasSameStatus = cardState && cardState.status === action.type;
-    if (hasSameStatus) {
-      return state;
-    } else {
-      const nextState = cardReducerMap[action.type](cardState, action);
-      return { ...state, [action.url]: nextState };
-    }
+  if (!isCardAction(action)) {
+    return state;
   }
-  return state;
+  const cardState = state[action.url];
+
+  // Card may have reached the same state on account of multiple of the same
+  // URL being present in an editor session. E.g. page with N links to one resource.
+  const hasSameStatus = cardState && cardState.status === action.type;
+  if (hasSameStatus) {
+    return state;
+  } else {
+    const nextState = reduceToNextState(cardState, action);
+    return { ...state, [action.url]: nextState };
+  }
 };