+++ linking-platform/linking-common/src/actions.ts
@@ -6,6 +6,7 @@ import { APIError } from './errors';
 export const ACTION_PENDING = 'pending';
 export const ACTION_RESOLVING = 'resolving';
 export const ACTION_RESOLVED = 'resolved';
+export const ACTION_RELOADING = 'reloading';
 export const ACTION_ERROR = 'errored';
 export const ACTION_ERROR_FALLBACK = 'fallback';
 export const ACTION_PRELOAD = 'preload';