+++ dev-portal/rating/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/rating
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.22
 
 ### Patch Changes